feat/jobCards-- styling refactored
