+++ client/src/App.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import './App.scss';
 
 function App() {
